% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.rtf-package}
\alias{blockr.rtf}
\alias{blockr.rtf-package}
\title{blockr.rtf: Blocks for parsing RTF files}
\description{
Blocks for reading and parsing RTF files to ARD format.
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

}
\keyword{internal}
