% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{new_card_block}
\alias{new_card_block}
\title{Read and parse Topline demo RTF files to CARD}
\usage{
new_card_block(
  file = character(),
  directory = blockr_option("topline_dir", system.file("extdata", "examples", package =
    "artful")),
  ...
)
}
\arguments{
\item{directory}{Directory where corresponding table is located}

\item{...}{Forwarded to \code{\link[blockr.core:new_block]{blockr.core::new_block()}}}

\item{table}{Table name}
}
\description{
This block allows to make read and parse specific RTF files shared as part
of an example topline slide deck.
}
