% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline.R
\name{new_topline_block}
\alias{new_topline_block}
\title{Read and parse Topline demo RTF files}
\usage{
new_topline_block(
  table = character(),
  directory = blockr_option("rtf_dir", system.file("extdata", "examples", package =
    "artful")),
  ...
)
}
\arguments{
\item{table}{Table name}

\item{directory}{Directory where corresponding table is located}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows to make read and parse specific RTF files shared as part
of an example topline slide deck.
}
