% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{new_rtf_block}
\alias{new_rtf_block}
\title{Read and parse RTF files}
\usage{
new_rtf_block(
  file = character(),
  directory = blockr_option("topline_dir", system.file("extdata", "examples", package =
    "artful")),
  parser = blockr_option("rtf_parser", "artful"),
  pivot = TRUE,
  ...
)
}
\arguments{
\item{file}{File name (in \code{\link[base:basename]{base::basename()}} sense)}

\item{directory}{Directory where \code{file} is located}

\item{parser}{RTF parser to use}

\item{pivot}{Logical, whether to pivot the grouping variables to long format.
Default is TRUE to maintain backward compatibility. When FALSE, the data
remains in wide format with grouping variables as columns.}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows to make read and parse an RTF file to an ARD formatted
\code{data.frame}.
}
