% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{new_rtf_to_df_block}
\alias{new_rtf_to_df_block}
\title{Read and parse an RTF file into a dataframe}
\usage{
new_rtf_to_df_block(
  file = character(),
  directory = blockr_option("topline_dir", system.file("extdata", "examples", package =
    "artful")),
  indentation_to_vars = TRUE,
  ...
)
}
\arguments{
\item{file}{File name (in \code{\link[base:basename]{base::basename()}} sense)}

\item{directory}{Directory where \code{file} is located}

\item{indentation_to_vars}{Logical; if \code{TRUE} (default), converts indentation
in the RTF table to explicit variables by calling
\code{\link[artful:indentation_to_variables]{artful::indentation_to_variables()}}.}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block reads and parses an RTF file into a \code{data.frame} object. The
function aims to return a copy of the original RTF table with as little data
preprocessing performed as possible (i.e., only essential steps such as
removing headers, footers, empty columns, etc., are performed). Unlike
\code{new_rtf_block()}, this function does not attempt to produce an ARD. Instead,
this step is off-loaded to other functions.
}
\examples{
\dontrun{
blockr.core::serve(new_rtf_to_df_block(file = "rt-ae-ae1.rtf"))
}
}
